### 2021-06-14 #3369 @whedon
[PRE REVIEW]: Building, Importing, and Exporting GEXF Graph Files with rgexf
**Submitting author:** @gvegayon (<a href="http://orcid.org/0000-0002-3171-0844">George Vega Yon</a>)
**Repository:** <a href="https://github.com/gvegayon/rgexf/" target ="_blank">https://github.com/gvegayon/rgexf/</a>
**Version:** 0.16.1
