### 2021-06-14 6dc94a1
Adding statement of need and fixing DOI https://github.com/openjournals/joss-reviews/issues/3369
