+++ paper/paper.md
@@ -23,42 +23,46 @@ affiliations:
 
 # Summary
 
-First introduced in 2012, the rgexf package for the R programming language was the first effort to make the Graph Exchange XML Format (@heymann2009gexf) specification available to the R world. With more than 500,000 downloads (https://cranlogs.r-pkg.org/), it is one of the most popular ways to incorporate GEXF files into the R programming language environment.
+First introduced in 2012, the **rgexf** package for the R programming language was the first effort to make the Graph Exchange XML Format [@heymann2009gexf] specification available to the R world. With more than 500,000 downloads[^cranlogs], it is one of the most popular ways to incorporate GEXF files into the R programming language environment.
 
-Developed by the Gephi Core Group (@bastian2009gephi), the GEXF specification is a flexible and widely used format to describe graphs. Although it has not been updated since 2009, the GEXF format has been introduced to several tools and programming environments. A few examples include:
+[^cranlogs]: According to the [https://cranlogs.r-pkg.org/](https://cranlogs.r-pkg.org/) website, as of June 14, 2021.
 
-- The python library `networkx` (@hagberg2008exploring) 
+Developed by the Gephi Core Group [@bastian2009gephi], the GEXF specification is a flexible and widely used format to describe graphs. Although it has not been updated since 2009, the GEXF format has been introduced to several tools and programming environments. A few examples include:
 
-- The stand-alone software `Cytoscape` (@smoot2011cytoscape)
+- The python library **networkx** [@hagberg2008exploring]
 
-- The JavaScript library `sigma.js` (https://simga.js)
+- The stand-alone software **Cytoscape** [@smoot2011cytoscape]
 
-- The java library `gexf4j` (https://github.com/francesco-ficarola/gexf4j) 
+- The JavaScript library **sigma.js** [https://simga.js](https://simga.js)
 
-- The JavaScript library `gexf-js` (https://github.com/raphv/gexf-js)
+- The java library **gexf4j** [https://github.com/francesco-ficarola/gexf4j)](https://github.com/francesco-ficarola/gexf4j) 
 
-Besides the rgexf package, other R packages provide functions that interact with GEXF files:
+- The JavaScript library **gexf-js** [https://github.com/raphv/gexf-js](https://github.com/raphv/gexf-js)
 
-- `sigmajs`: Interface to 'Sigma.js' Graph Visualization Library (@Coene2018)
+Besides the **rgexf** package, other R packages provide functions that interact with GEXF files:
 
-- `vkR`: Access to VK API via R (@vkR)
+- **sigmajs**: Interface to 'Sigma.js' Graph Visualization Library [@Coene2018]
 
-- `microeco`: Microbial Community Ecology Data Analysis (@microeco) 
+- **vkR**: Access to VK API via R [@vkR]
 
-- `netCoin`: Interactive Analytic Networks (@netCoin)
+- **microeco**: Microbial Community Ecology Data Analysis [@microeco]
 
-Nevertheless, the rgexf package continues to be the de-facto tool to interact with GEXF files in R.
+- **netCoin**: Interactive Analytic Networks [@netCoin]
+
+Nevertheless, the **rgexf** package continues to be the de-facto tool to interact with GEXF files in R.
+
+# Statement of Need
+
+This R package has been serving the scientific community for many years now. Scientists and data analysts across the board have been using **rgexf** to enhance their analyses by smoothly moving between **R** and other applications used for graph visualization. Some concrete examples include gene networks [@Starr2017;@Kauffman2018], interactions among species [@Leclerc2018], and social networks [@Alsaedi2016].
 
 # Features
 
-Beyond reading and writing GEXF files from within R, the rgexf R package has various other features that can help to create beautiful network visualizations, in particular:
[…]
