+++ paper/paper.bib
@@ -31,15 +31,22 @@
 }
 
 @article{smoot2011cytoscape,
-  title={Cytoscape 2.8: new features for data integration and network visualization},
-  author={Smoot, Michael E and Ono, Keiichiro and Ruscheinski, Johannes and Wang, Peng-Liang and Ideker, Trey},
-  journal={Bioinformatics},
-  volume={27},
-  number={3},
-  pages={431--432},
-  year={2011},
-  publisher={Oxford University Press}
+    author = {Smoot, Michael E. and Ono, Keiichiro and Ruscheinski, Johannes and Wang, Peng-Liang and Ideker, Trey},
+    title = "{Cytoscape 2.8: new features for data integration and network visualization}",
+    journal = {Bioinformatics},
+    volume = {27},
+    number = {3},
+    pages = {431-432},
+    year = {2010},
+    month = {12},
+    abstract = "{Summary: Cytoscape is a popular bioinformatics package for biological network visualization and data integration. Version 2.8 introduces two powerful new features—Custom Node Graphics and Attribute Equations—which can be used jointly to greatly enhance Cytoscape's data integration and visualization capabilities. Custom Node Graphics allow an image to be projected onto a node, including images generated dynamically or at remote locations. Attribute Equations provide Cytoscape with spreadsheet-like functionality in which the value of an attribute is computed dynamically as a function of other attributes and network properties.Availability and implementation: Cytoscape is a desktop Java application released under the Library Gnu Public License (LGPL). Binary install bundles and source code for Cytoscape 2.8 are available for download from http://cytoscape.org.Contact:msmoot@ucsd.edu}",
+    issn = {1367-4803},
+    doi = {10.1093/bioinformatics/btq675},
+    url = {https://doi.org/10.1093/bioinformatics/btq675},
+    eprint = {https://academic.oup.com/bioinformatics/article-pdf/27/3/431/917014/btq675.pdf},
 }
+
+
 
 @article{Coene2018,
   doi = {10.21105/joss.00814},
