### 2021-06-30 #3369 @fabian-s
@jonjoncardoso @vc1492a @jankatins would any of you be willing to review this submission for JOSS?

Thank you for considering this!
