### 2021-06-30 #3369 @vc1492a
@fabian-s thanks for reaching out! I am afraid my R language experience these days is pretty limited (although it used to be much better), and think another reviewer may be better positioned to do a thorough review on this one. Good luck!
