### 2021-07-01 #3369 @fabian-s
@briatte would you be willing to review this submission for JOSS?

Thank you for considering this!
