### 2021-07-07 #3456 @whedon
[REVIEW]: Building, Importing, and Exporting GEXF Graph Files with rgexf
**Submitting author:** @gvegayon (<a href="http://orcid.org/0000-0002-3171-0844">George Vega Yon</a>)
**Repository:** <a href="https://github.com/gvegayon/rgexf/" target ="_blank">https://github.com/gvegayon/rgexf/</a>
**Version:** v0.16.2

- [x] #3520
