### 2021-07-22 #3456 @jonjoncardoso
The only thing I missed was a `CONTRIBUTING.md` file or an equivalent section on the README to highlight how people could contribute to the code so as to the journal's documentation checklist:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Other than that, all good from my side. I've managed to install, follow the examples and generate some .gexf files myself.

I have used rgexf in the past for a netviz project and I remember it being very useful.
