### 2021-07-25 #3456 @corneliusfritz
I also came to the same conclusion. There seem to be 16 open issues that are all not entirely solved, as far as I can see. I only have some minor additional remarks: Some tests/examples not nested in a demo, but the actual "Examples" Section would be helpful; also, the break commands should be placed so that each segment of code is one line to enable a faster learning curve and not necessitate scrolling through code all the time. Besides, the documentation of dynamic edges is not completely clear to me.  For each edge, we need to provide some time where the edge starts and where it ends. Does the first column always need to be the time when it starts? If so, how are the examples interpreted, where the time indicated in the second row is before the time of the first row?  And are the NAs marking some infinite end or beginning? Finally, what is the difference between a spell and an edge?

So far, I have always used CSV files to export my network data to Gephi, which works as well but going through the examples and trying out some things with my data, I believe that the package does some excellent and helpful work in that I will use it in future applications.  In particular, the plot functionality is a nice add-on.
