### 2021-07-27 gvegayon/rgexf#43 @gvegayon
JOSS Review
This issue contains the comments and suggestions raised by @jonjoncardoso and @corneliusfritz in https://github.com/openjournals/joss-reviews/issues/3456:

- [x] Add a `code_of_conduct.md` file ([ref](https://github.com/openjournals/joss-reviews/issues/3456#issuecomment-884884471))
- [x] Review the "Examples" section: Breaks ([ref](https://github.com/openjournals/joss-reviews/issues/3456#issuecomment-886240844)).
- [x] Improve the doc on dynamic edges.
- [x] Clarify the difference between spell and edge.
- [x] Credit where credit is due: Add both reviewers to the list of authors (😃)
