### 2021-08-02 e0945ca
Adding CODE_OF_CONDUCT.md (#43)
