### 2021-08-07 02b828e
Working on examples and credit (#43)
