+++ R/methods.R
@@ -1,6 +1,6 @@
 #' @name gexf-methods
 #' @export
-print.gexf <- function(x, file=NA, replace=F, ...) {
+print.gexf <- function(x, file = NA, replace = F, ...) {
 ################################################################################
 # Printing method
 ################################################################################
@@ -13,16 +13,21 @@ print.gexf <- function(x, file=NA, replace=F, ...) {
 #' @name gexf-methods
 #' @export
 summary.gexf <- function(object, ...) {
+
   ################################################################################
   # Printing method
   ################################################################################
-  result <- list("N of nodes"=NROW(object$nodes), 
-                 "N of edges"=NROW(object$edges),
-                 "Node Attrs"=utils::head(object$atts.definitions$node.att),
-                 "Edge Attrs"=utils::head(object$atts.definitions$edge.att))
+  result <- list(
+    "N of nodes"=NROW(object$nodes), 
+    "N of edges"=NROW(object$edges),
+    "Node Attrs"=utils::head(object$atts.definitions$nodes),
+    "Edge Attrs"=utils::head(object$atts.definitions$edges)
+  )
+  
   #class(result) <- "table"
   cat("GEXF graph object\n")
   result
+
 }
 
 build.and.validate.gexf <- function(
