+++ demo/gexfattributes.R
@@ -37,7 +37,7 @@ edge.att
 #
 # before the last closing 
 # parenthesis in the following function
-
+pause()
 gexf(nodes=people, edges=relations, nodesAtt=node.att, edgesAtt=edge.att)
 
 ################################################################################
