+++ tests/testthat/test-examples.r
@@ -1,3 +1,0 @@
-# context("Examples")
-# 
-# test_examples()
