+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(rgexf)
-
-test_check("rgexf")
