+++ docker/Makefile
@@ -1,6 +1,4 @@
 build:
-	docker build -t uscbiostats/aphylo:release -f Dockerfile . && \
-		docker build -t uscbiostats/aphylo:devel -f Dockerfile.dev .
+	docker build -t gvegayon/rgexf:release -f Dockerfile . 
 push: build
-	docker push uscbiostats/aphylo:release && \
-		docker push uscbiostats/aphylo:devel
+	docker push gvegayon/rgexf:release 
