### 2021-08-10 5be3fbb
Adding more examples, a head method
