+++ man/followers.Rd
@@ -4,7 +4,9 @@
 \name{followers}
 \alias{followers}
 \title{Edge list with attributes}
-\format{A data frame containing 6065 observations.}
+\format{
+A data frame containing 6065 observations.
+}
 \source{
 Fabrega and Paredes (2012): \dQuote{La politica en 140 caracteres}
 en Intermedios: medios de comunicacion y democracia en Chile. Ediciones UDP
