+++ docker/Dockerfile
@@ -1,9 +1,11 @@
-FROM rocker/r-base:latest
+FROM rocker/drd
 
 RUN apt-get update && \
 	apt-get install -y --no-install-recommends \
-	r-cran-xml r-cran-igraph r-cran-rmarkdown r-cran-knitr \
-	r-cran-covr r-cran-tinytest
+	libssl-dev
+
+RUN Rscript -e 'utils::install.packages(c("XML", "igraph", "rmarkdown", "knitr", \
+	"covr", "tinytest")'
 
 RUN Rscript -e 'utils::install.packages("servr")'
 
