+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(head,gexf)
 S3method(plot,gexf)
 S3method(print,gexf)
 S3method(summary,gexf)

[…]
