### 2021-08-10 9b52f78
Improving documentation + more examples (#43)
