+++ R/plot.gexf.R
@@ -1,8 +1,0 @@
-#' #' @name gexf-methods
-#' #' @export
-#' plot.gexf <- function(x, EdgeType=NULL, output.dir=NULL,...){  
-#'   # WIP
-#' 
-#'   .Defunct("Gephi for visualizing")  
-#' }
-#' 
