+++ R/gexf.R
@@ -1,0 +1,664 @@
+#' Creates an object of class `gexf`
+#' 
+#' Takes a `node` matrix (or dataframe) and an
+#' `edge` matrix (or dataframe) and creates a `gexf` object
+#' containing a data-frame representation and a gexf representation of a graph.
+#' 
+#' @details
+#' Just like `nodesVizAtt` and `edgesVizAtt`, `nodesAtt` and
+#' `edgesAtt` must have the same number of rows as nodes and edges,
+#' respectively. Using data frames is necessary as in this way data types are
+#' preserved.
+#' 
+#' `nodesVizAtt` and `edgesVizAtt` allow using visual attributes such
+#' as color, position (nodes only), size (nodes only), thickness (edges only)
+#' shape and image (nodes only).  \itemize{ \item Color is defined by the RGBA
+#' color model, thus for every node/edge the color should be specified through
+#' a data-frame with columns *r* (red), *g* (green), *b* (blue)
+#' with integers between 0 and 256 and a last column with *alpha* values
+#' as a float between 0.0 and 1.0.  \item Position, for every node, it is a
+#' three-column data-frame including *x*, *y* and *z*
+#' coordinates. The three components must be float.  \item Size as a numeric
+#' colvector (float values).  \item Thickness (see size).  \item Node Shape
+#' (string), currently unsupported by Gephi, can take the values of
+#' *disk*, *square*, *triangle*, *diamond* and
+#' *image*.  \item Edge Shape (string), currently unsupported by Gephi, can
+#' take the values of *solid*, *dotted*, *dashed* and
+#' *double*.  \item Image (string), currently unsupported by Gephi,
+#' consists on a vector of strings representing URIs.  }
+#' 
+#' `nodeDynamic` and `edgeDynamic` allow to draw dynamic graphs. It
+#' should contain two columns *start* and *end*, both allowing
+#' `NA` value. It can be use jointly with `tFormat` which by default
+#' is set as \dQuote{double}. Currently accepted time formats are: 
+#' 
+#' - Integer or double.
+#' - International standard *date* yyyy-mm-dd.
+#' - dateTime W3 XSD (http://www.w3.org/TR/xmlschema-2/#dateTime).
+#' 
+#' `NA` values in the first column are filled with the min of
+#' `c(nodeDynamic, edgeDynamic)`, whereas if in the second column
+#' is replaces with the max.
+#' 
+#' More complex time sequences like present/absent nodes and edges
+#' can be added with [add.node.spell] and [add.edge.spell] respectively.
+#' 
+#' @param nodes A two-column data-frame or matrix of \dQuote{id}s and
+#' \dQuote{label}s representing nodes.
+#' @param edges A two-column data-frame or matrix containing \dQuote{source}
+#' and \dQuote{target} for each edge. Source and target values are based on the
+#' nodes ids.
+#' @param edgesId A one-column data-frame, matrix or vector.
+#' @param edgesLabel A one-column data-frame, matrix or vector.
+#' @param edgesAtt A data-frame with one or more columns representing edges'
+#' attributes.
+#' @param edgesWeight A numeric vector containing edges' weights.
+#' @param edgesVizAtt List of three or less viz attributes such as color, size
[…]
