+++ R/manipulation.R
@@ -1,0 +1,593 @@
+#' Adding and removing nodes/edges from `gexf` objects
+#' 
+#' Manipulates `gexf` objects adding and removing nodes and edges from
+#' both, its dataframe representation and its XML representation.
+#' 
+#' `new.gexf.graph` Creates a new `gexf` empty object (0 nodes 0
+#' edges).
+#' 
+#' `add.gexf.node` and `add.gexf.edge` allow adding nodes and edges
+#' to a `gexf` object (graph) one at a time. `rm.gexf.node` and
+#' `rm.gexf.edges` remove nodes and edges respectively.
+#' 
+#' In the case of `rm.gexf.node`, by default every edge linked to the node
+#' that is been removed will also be removed (`rm.edges = TRUE`).
+#' 
+#' @aliases add.gexf.node add.gexf.edge rm.gexf.node rm.gexf.edge
+#' add.node.spell add.edge.spell
+#' @param graph A gexf-class object.
+#' @param id A node/edge id (normally numeric value).
+#' @param label A node/edge label.
+#' @param type Type of connection (edge).
+#' @param number Index number(s) of a single or a group of nodes or edges.
+#' @param weight Edge weight.
+#' @param vizAtt A list of node/edge viz attributes (see
+#' [write.gexf()]).
+#' @param atts List of attributes, currently ignored.
+#' @param source Source node's id.
+#' @param target Target node's id.
+#' @param start Starting time period
+#' @param end Ending time period
+#' @param rm.edges Whether to remove or not existing edges.
+#' @param digits Integer. Number of decimals to keep for nodes/edges sizes. See
+#' [print.default()]
+#' @return A `gexf` object (see [write.gexf()]).
+#' @author George Vega Yon 
+#' 
+#' Jorge Fabrega Lacoa 
+#' @references The GEXF project website: http://gexf.net
+#' @keywords manip
+#' @examples
+#'
+#' if (interactive()) {
+#'   demo(gexfbuildfromscratch)
+#' }
+#' 
+#' @export add.gexf.node
+add.gexf.node <- function(
+  graph, 
+  id     = NA, 
+  label  = NA, 
+  start  = NULL, 
+  end    = NULL,
+  vizAtt = list(
+    color    = NULL,
+    position = NULL,
+    size     = NULL,
+    shape    = NULL,
+    image    = NULL
+  ),
+  atts   = NULL
+  ) 
+  {
+
+  # Checks the class
+  if (!inherits(graph,"gexf")) stop("-graph- is not of -gexf- class.")
+
+  # Parses the graph file
+  graph$graph <- XML::xmlTreeParse(graph$graph, encoding="UTF-8")
+  
+  # Gets the number of nodes
+  n <- length(graph$graph$doc$children$gexf[["graph"]][["nodes"]])
+  
+  node <- XML::xmlNode("node", attrs=c(id=id, label=label, start=start, end=end))
+  
+  # Adds the atts
+#   if (length(atts)) {
+#     atts.node <- xmlNode("attvalues")
+#     for (i in 1:length(atts)) {
+#       atts.node <- 
+#         addChildren(
+#           atts.node, 
+#           xmlNode("attvalue", attrs=c("for"=names(atts)[i], value=atts[[i]])))
+#       
+#       # Checking atts definition
+#       if (!length(graph$atts.definitions$node)) {
+#         graph$atts.definitions$node <-
+#           data.frame(
+#             id=names(atts)[i], 
+#             title=atts[[i]], 
+#             type=.parseDataTypes(atts[[i]])
[…]
