+++ man/plot.gexf.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gexf-js.R
+% Please edit documentation in R/plot.R
 \name{gexf_js_config}
 \alias{gexf_js_config}
 \alias{plot.gexf}
