+++ README.md
@@ -106,22 +106,22 @@ write.gexf(people, relations)
     ##     <nodes>
     ##       <node id="1" label="juan">
     ##         <viz:color r="255" g="99" b="71" a="1"/>
-    ##         <viz:position x="84.4622068841045" y="-100" z="0"/>
+    ##         <viz:position x="100" y="51.0986544575574" z="0"/>
     ##         <viz:size value="10"/>
     ##       </node>
     ##       <node id="2" label="pedro">
     ##         <viz:color r="255" g="99" b="71" a="1"/>
-    ##         <viz:position x="100" y="100" z="0"/>
+    ##         <viz:position x="-100" y="100" z="0"/>
     ##         <viz:size value="10"/>
     ##       </node>
     ##       <node id="3" label="matthew">
     ##         <viz:color r="255" g="99" b="71" a="1"/>
-    ##         <viz:position x="13.6250760243742" y="7.95768386709403" z="0"/>
+    ##         <viz:position x="-14.2129423502891" y="83.9680923396099" z="0"/>
     ##         <viz:size value="10"/>
     ##       </node>
     ##       <node id="4" label="carlos">
     ##         <viz:color r="255" g="99" b="71" a="1"/>
-    ##         <viz:position x="-100" y="78.79797591046" z="0"/>
+    ##         <viz:position x="66.396171662381" y="-100" z="0"/>
     ##         <viz:size value="10"/>
     ##       </node>
     ##     </nodes>
@@ -163,26 +163,26 @@ write.gexf(people, relations, nodeDynamic=time)
     ##     <description>A GEXF file written in R with "rgexf"</description>
     ##     <keywords>GEXF, NodosChile, R, rgexf, Gephi</keywords>
     ##   </meta>
-    ##   <graph mode="dynamic" start="10" end="2" timeformat="double" defaultedgetype="undirected">
+    ##   <graph mode="dynamic" start="2" end="13" timeformat="double" defaultedgetype="undirected">
     ##     <nodes>
     ##       <node id="1" label="juan" start="10" end="12">
     ##         <viz:color r="255" g="99" b="71" a="1"/>
-    ##         <viz:position x="69.6983538631002" y="-69.8983985366003" z="0"/>
+    ##         <viz:position x="100" y="-66.7395710811224" z="0"/>
     ##         <viz:size value="10"/>
     ##       </node>
-    ##       <node id="2" label="pedro" start="13" end="2">
+    ##       <node id="2" label="pedro" start="13" end="13">
     ##         <viz:color r="255" g="99" b="71" a="1"/>
-    ##         <viz:position x="100" y="-100" z="0"/>
+    ##         <viz:position x="13.4431787806887" y="61.4550911958847" z="0"/>
     ##         <viz:size value="10"/>
     ##       </node>
-    ##       <node id="3" label="matthew" start="2" end="2">
-    ##         <viz:color r="255" g="99" b="71" a="1"/>
-    ##         <viz:position x="-73.6968091913099" y="54.8684641665018" z="0"/>
-    ##         <viz:size value="10"/>
-    ##       </node>
-    ##       <node id="4" label="carlos" start="2" end="2">
+    ##       <node id="3" label="matthew" start="2" end="13">
     ##         <viz:color r="255" g="99" b="71" a="1"/>
     ##         <viz:position x="-100" y="100" z="0"/>
+    ##         <viz:size value="10"/>
+    ##       </node>
+    ##       <node id="4" label="carlos" start="2" end="13">
+    ##         <viz:color r="255" g="99" b="71" a="1"/>
+    ##         <viz:position x="-79.4152158871745" y="-100" z="0"/>
     ##         <viz:size value="10"/>
     ##       </node>
     ##     </nodes>
[…]
