+++ DESCRIPTION
@@ -2,8 +2,8 @@ Package: rgexf
 Type: Package
 Encoding: UTF-8
 Title: Build, Import and Export GEXF Graph Files
-Version: 0.16.1
-Date: 2021-06-12
+Version: 0.16.2
+Date: 2021-08-10
 Authors@R: c(
   person("George", "Vega Yon", email="g.vegayon@gmail.com", role=c("aut", "cre"),
   comment = structure("0000-0002-3171-0844", .Names = "ORCID")),
