+++ .github/workflows/website.yml
@@ -16,7 +16,9 @@ jobs:
         run: docker pull ${cntr}
 
       - name: Building website
-        run: docker run --rm -i -v ${PWD}:/home/mnt -w /home/mnt ${cntr} Rscript -e 'install.packages(c("igraph", "servr", "tinytest"));system("R CMD INSTALL .");pkgdown::build_site(install=FALSE)'
+        run: docker run --rm -i -v ${PWD}:/home/mnt -w /home/mnt ${cntr} apt-get install --no-install-recommends -y \
+          r-cran-igraph r-cran-xml; \
+          Rscript -e 'install.packages(c("servr", "tinytest"));system("R CMD INSTALL .");pkgdown::build_site(install=FALSE)'
 
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
