+++ .github/workflows/website.yml
@@ -3,7 +3,7 @@ name: rgexf website
 on: [push, pull_request]
 
 env:
-  cntr: rocker/tidyverse:latest
+  cntr: gvegayon/rgexf:latest
 
 jobs:
   build:
@@ -16,7 +16,7 @@ jobs:
         run: docker pull ${cntr}
 
       - name: Building website
-        run: docker run --rm -i -v ${PWD}:/home/mnt -w /home/mnt ${cntr} Rscript -e 'install.packages(c("igraph", "servr", "tinytest"));system("R CMD INSTALL .");pkgdown::build_site(install=FALSE)'
+        run: docker run --rm -i -v ${PWD}:/home/mnt -w /home/mnt ${cntr} Rscript -e 'pkgdown::build_site()'
 
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
