+++ pkgdown.yml
@@ -1,7 +1,7 @@
-pandoc: '2.2'
-pkgdown: 1.5.1
+pandoc: 2.9.2.1
+pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   rgexf: rgexf.html
-last_built: 2020-06-15T17:07Z
+last_built: 2021-08-10T18:20Z
 
