+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   rgexf: rgexf.html
-last_built: 2021-08-10T18:20Z
+last_built: 2021-08-10T18:26Z
 
