+++ README.Rmd
@@ -148,6 +148,12 @@ par(op)
 
 # Code of Conduct
   
+We welcome contributions to `rgexf`. Whether reporting a bug, starting a
+discussion by asking a question, or proposing/requesting a new feature,
+please go by creating a new issue
+[here](https://github.com/gvegayon/rgexf/issues) so that we can talk
+about it.
+
 Please note that the rgexf project is released with a [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms
 
 
