+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   rgexf: rgexf.html
-last_built: 2021-08-10T19:27Z
+last_built: 2021-08-10T19:30Z
 
