### 2021-08-10 a2beb32
Finalizing spell example
