+++ man/add.gexf.node.Rd
@@ -147,6 +147,7 @@ g <- add.node.spell(g, 0, start = 3)
 g <- add.edge.spell(g, 1, start = 1, end = 2)
 g <- add.edge.spell(g, 1, start = 3.5)
 
+g
 
 }
 \references{
