+++ R/read.gexf.R
@@ -10,7 +10,7 @@
 #' @author George Vega Yon 
 #' 
 #' Jorge Fabrega Lacoa 
-#' @references The GEXF project website: http://gexf.net
+#' @references The GEXF project website: https://gephi.org/gexf/format/
 #' @keywords IO
 #' @examples
 #' 
