+++ README.md
@@ -10,7 +10,6 @@ website](https://github.com/gvegayon/rgexf/actions/workflows/website.yml/badge.s
 Status](https://ci.appveyor.com/api/projects/status/github/gvegayon/rgexf?branch=master&svg=true)](https://ci.appveyor.com/project/gvegayon/rgexf)
 [![Coverage
 Status](https://img.shields.io/codecov/c/github/gvegayon/rgexf/master.svg)](https://codecov.io/github/gvegayon/rgexf?branch=master)
-[![lifecycle](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://www.tidyverse.org/lifecycle/#stable)
 
 # rgexf: Build, Import and Export GEXF Graph Files <img src="man/figures/logo.svg" align="right" height="200"/>
 
@@ -35,33 +34,33 @@ the igraph package.
 
 ## New features and changes
 
-  - We now have a hex sticker\!
+-   We now have a hex sticker!
 
-  - Users can now cite properly `rgexf` with JOSS (see
+-   Users can now cite properly `rgexf` with JOSS (see
     `citation(package="rgexf")`).
 
-  - Added a few extra breaks across the examples (suggested by
+-   Added a few extra breaks across the examples (suggested by
     @corneliusfritz).
 
-  - Improved documentation regarding spells and dynamic graphs.
+-   Improved documentation regarding spells and dynamic graphs.
 
-  - New `head()` function allows a glimpse of the `n` first nodes/edges.
+-   New `head()` function allows a glimpse of the `n` first nodes/edges.
 
 ## Bug fixes
 
-  - Passing colors with four values (alpha) no longer fails during
+-   Passing colors with four values (alpha) no longer fails during
     checks. (reported by @IsabelFE).
 
-  - The summary function prints nodes’ attributes as expected.
+-   The summary function prints nodes’ attributes as expected.
 
-  - Hex colors now work (\#41 reported by @milnus).
+-   Hex colors now work (\#41 reported by @milnus).
 
-  - `gexf.to.igraph` correctly processes edge attributes (\#38 reported
+-   `gexf.to.igraph` correctly processes edge attributes (\#38 reported
     by @balachia).
 
-  - Time range is now correctly computed (\#19).
+-   Time range is now correctly computed (\#19).
 
-  - Non-integer ids were incorrectly processed when reading GEXF files.
+-   Non-integer ids were incorrectly processed when reading GEXF files.
 
 More in the [NEWS.md](NEWS.md) file.
 
@@ -118,7 +117,7 @@ write.gexf(people, relations)
 
     ## <?xml version="1.0" encoding="UTF-8"?>
     ## <gexf xmlns="http://www.gexf.net/1.3" xmlns:viz="http://www.gexf.net/1.3/viz" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.gexf.net/1.3 http://www.gexf.net/1.3/gexf.xsd" version="1.3">
-    ##   <meta lastmodifieddate="2021-08-10">
+    ##   <meta lastmodifieddate="2021-08-11">
     ##     <creator>NodosChile</creator>
     ##     <description>A GEXF file written in R with "rgexf"</description>
     ##     <keywords>GEXF, NodosChile, R, rgexf, Gephi</keywords>
@@ -127,22 +126,22 @@ write.gexf(people, relations)
     ##     <nodes>
     ##       <node id="1" label="juan">
     ##         <viz:color r="255" g="99" b="71" a="1"/>
-    ##         <viz:position x="-100" y="-83.9578418284886" z="0"/>
[…]
