+++ man/read.gexf.Rd
@@ -26,7 +26,7 @@ By the time attributes and viz-attributes aren't supported.
 
 }
 \references{
-The GEXF project website: http://gexf.net
+The GEXF project website: https://gephi.org/gexf/format/
 }
 \author{
 George Vega Yon
