+++ man/add.gexf.node.Rd
@@ -151,7 +151,7 @@ g
 
 }
 \references{
-The GEXF project website: http://gexf.net
+The GEXF project website: https://gephi.org/gexf/format/
 }
 \author{
 George Vega Yon
