+++ man/rgexf-package.Rd
@@ -50,7 +50,7 @@ if (interactive()) {
 \references{
 \itemize{ \item rgexf project site:
 \url{https://github.com/gvegayon/rgexf} \item Gephi project site:
-\url{https://gephi.org/} \item GEXF project site: \url{http://gexf.net/}
+\url{https://gephi.org/} \item GEXF project site: \url{https://gephi.org/gexf/format//}
 \item gexf-js project website: \url{https://github.com/raphv/gexf-js}
 \item Sigmasj project site: \url{http://sigmajs.org/}
 }
