+++ man/new.gexf.graph.Rd
@@ -31,7 +31,7 @@ if (interactive()) {
 
 }
 \references{
-The GEXF project website: http://gexf.net
+The GEXF project website: https://gephi.org/gexf/format/
 }
 \author{
 George Vega Yon
