### 2021-08-11 #3456 @gvegayon
- [x] Tag: [v0.16.2](https://github.com/gvegayon/rgexf/tree/v0.16.2)
- [x] Zenodo archive: [here](https://zenodo.org/record/5182709)
- [x] Title matches paper (and right list of authors)
- [x] DOI: [10.5281/zenodo.5182709](https://doi.org/10.5281/zenodo.5182709)
